feat: add SeriesManager to manage series nodes
