+++ svg-time-series/bench/renderPaths.bench.ts
@@ -4,12 +4,14 @@
 import { bench, describe } from "vitest";
 import { select } from "d3-selection";
 import { SeriesRenderer } from "../src/chart/seriesRenderer.ts";
+import { SeriesManager } from "../src/chart/series.ts";
 import { sizes, datasets } from "./timeSeriesData.ts";
 
 describe("renderPaths performance", () => {
   const svg = document.createElementNS("http://www.w3.org/2000/svg", "svg");
   const renderer = new SeriesRenderer();
-  renderer.init(select(svg), 2, [0, 1]);
+  const manager = new SeriesManager(select(svg), 2, [0, 1]);
+  renderer.series = manager.series;
 
   sizes.forEach((size, idx) => {
     const data = datasets[idx];
