+++ svg-time-series/src/chart/render/utils.ts
@@ -30,3 +30,14 @@ export function updateScaleX(
   const bTimeVisible = bIndexVisible.transformWith(transform);
   x.domain(bTimeVisible.toArr());
 }
+
+export function createSeriesNodes(
+  svg: Selection<SVGSVGElement, unknown, HTMLElement, unknown>,
+): { view: SVGGElement; path: SVGPathElement } {
+  const view = svg.append("g").attr("class", "view");
+  const path = view.append<SVGPathElement>("path");
+  return {
+    view: view.node() as SVGGElement,
+    path: path.node() as SVGPathElement,
+  };
+}
