+++ svg-time-series/src/chart/series.ts
@@ -1,0 +1,28 @@
+import { Selection } from "d3-selection";
+import { line, type Line } from "d3-shape";
+
+import type { Series } from "./render.ts";
+import { createSeriesNodes } from "./render/utils.ts";
+
+export class SeriesManager {
+  public readonly series: Series[] = [];
+
+  constructor(
+    svg: Selection<SVGSVGElement, unknown, HTMLElement, unknown>,
+    seriesCount: number,
+    seriesAxes: number[],
+  ) {
+    for (let i = 0; i < seriesCount; i++) {
+      const { view, path } = createSeriesNodes(svg);
+      const axisIdx = seriesAxes[i] ?? 0;
+      this.series.push({ axisIdx, view, path, line: this.createLine(i) });
+    }
+  }
+
+  private createLine(seriesIdx: number): Line<number[]> {
+    return line<number[]>()
+      .defined((d) => !(isNaN(d[seriesIdx]) || d[seriesIdx] == null))
+      .x((_, i) => i)
+      .y((d) => d[seriesIdx] as number);
+  }
+}
